+++ src/cljs/cxengage_javascript_sdk/interaction_management.cljs
@@ -267,6 +267,10 @@
   (p/publish {:topics (p/get-topic :resource-added)
               :response message}))
 
+(defn handle-resource-removed [message]
+  (p/publish {:topics (p/get-topic :resource-removed)
+              :response message}))
+
 (defn handle-screen-pop [message]
   (let [{:keys [pop-uri pop-type interaction-id]} message]
     (when (and pop-uri (or (= pop-type "external-url") (= pop-type "url")))
@@ -320,6 +324,7 @@
                       :SESSION/CHANGE_STATE_RESPONSE handle-resource-state-change
                       :SESSION/START_SESSION_RESPONSE handle-session-start
                       :INTERACTIONS/RESOURCE_ADDED handle-resource-added
+                      :INTERACTIONS/RESOURCE_REMOVED handle-resource-removed
                       nil)]
     (when (and (get message :action-id)
                (not= (get message :interaction-id) "00000000-0000-0000-0000-000000000000")
@@ -363,6 +368,7 @@
                                        "recording-start" :INTERACTIONS/RECORDING_START_RECEIVED
                                        "recording-stop" :INTERACTIONS/RECORDING_STOP_RECEIVED
                                        "resource-added" :INTERACTIONS/RESOURCE_ADDED
+                                       "resource-removed" :INTERACTIONS/RESOURCE_REMOVED
                                        "transfer-connected" :INTERACTIONS/TRANSFER_CONNECTED_RECEIVED
                                        :INTERACTIONS/GENERIC_AGENT_NOTIFICATION)]
       (merge {:sdk-msg-type inferred-notification-type} message))))